Shuffles components into reasonable directory structure.
