+++ src/lib/Scenes/Home/__tests__/SectionTitle-tests.tsx
