+++ src/lib/Scenes/Sales/Components/LotsByFollowedArtists.tsx
