+++ src/lib/Scenes/Sales/Components/SaleListItem.tsx
