+++ src/lib/Scenes/Sales/Components/__tests__/LotsByFollowedArtists-tests.tsx
